Move to domain specific folder structure
