+++ client/my-sites/checkout/checkout-thank-you/domains/domain-mapping-thank-you.jsx
