+++ client/my-sites/checkout/checkout-thank-you/index.jsx
@@ -73,8 +73,8 @@ import BloggerPlanDetails from './blogger-plan-details';
 import BusinessPlanDetails from './business-plan-details';
 import ChargebackDetails from './chargeback-details';
 import DomainMappingDetails from './domain-mapping-details';
-import DomainMappingThankYou from './domain-mapping-thank-you';
 import DomainRegistrationDetails from './domain-registration-details';
+import DomainMappingThankYou from './domains/domain-mapping-thank-you';
 import EcommercePlanDetails from './ecommerce-plan-details';
 import FailedPurchaseDetails from './failed-purchase-details';
 import CheckoutThankYouFeaturesHeader from './features-header';
